fix test

+++ server/routes/post-routes.js
@@ -77,7 +77,12 @@ try {
     list: postList,
     filename: __dirname.replace('/server/routes', '') + '/src/views/index.ejs'
   });
-  fs.writeFileSync(`public/index.html`, html, 'utf8');
+  if (!fs.existsSync(`public`)) {
+    fs.mkdirSync(`public`);
+    fs.writeFileSync(`public/index.html`, html, 'utf8');
+  } else {
+    fs.writeFileSync(`public/index.html`, html, 'utf8');
+  }
 } catch (e) {
   console.log(e);
 }
@@ -90,7 +95,12 @@ try {
     list: postList,
     filename: __dirname.replace('/server/routes', '') + '/src/views/about.ejs'
   });
-  fs.writeFileSync(`public/about.html`, html, 'utf8');
+  if (!fs.existsSync(`public`)) {
+    fs.mkdirSync(`public`);
+    fs.writeFileSync(`public/about.html`, html, 'utf8');
+  } else {
+    fs.writeFileSync(`public/about.html`, html, 'utf8');
+  }
 } catch (e) {
   console.log(e);
 }
@@ -105,11 +115,16 @@ postList.map((post, index) => {
       body: post.body,
       filename: __dirname.replace('/server/routes', '') + '/src/views/static.ejs'
     });
-    fs.writeFileSync(`public/post/${post.filename}.html`, html, 'utf8');
+    if (!fs.existsSync(`public/post/`)) {
+      fs.mkdirSync(`public/post/`);
+      fs.writeFileSync(`public/post/${post.filename}.html`, html, 'utf8');
+    } else {
+      fs.writeFileSync(`public/post/${post.filename}.html`, html, 'utf8');
+    }
   } catch (e) {
     console.log(e);
   }
-});
+}); 
 
 //create tag post files
 navLinks.map((tag, index) => {
@@ -140,42 +155,21 @@ navLinks.map((tag, index) => {
 router.get('/', (req, res) => {
   console.log(req.params);
   res.sendFile(ROOT_DIR + `/index.html`);
-  // res.render('index', {
-  //   nav: sortedNav,
-  //   list: postList
-  // });
 });
 
 //about page
 router.get('/about', (req, res) => {
   res.sendFile(ROOT_DIR + `/about.html`);
-  // res.render('index', {
-  //   nav: sortedNav,
-  //   list: postList
-  // });
 });
 
 //render list from tags
 router.get('/:tag', (req, res) => {
   res.sendFile(ROOT_DIR + `/${req.params.tag}/index.html`);
-  // if (tagPost[req.params.tag]) {
-  //   res.render('index', {
-  //     title: "sup dunny",
-  //     nav: sortedNav,
-  //     list: tagPost[req.params.tag]
-  //   });
-  // }
 });
 
 //render single post
 router.get('/post/:post', (req, res) => {
   res.sendFile(ROOT_DIR + `/post/${req.params.post}.html`);
-  // const post = postData[req.params.post + '.md'];
-  // res.render('post', {
-  //   nav: sortedNav,
-  //   post: post,
-  //   body: post.body
-  // });
 });
 
 module.exports = router;
